Have the api recieving data now. It is not sending it thru the paylaod correctly yet though. I made a file DisplayedData.jsx. The is fetching is true but not switched off yet.
