+++ react-redux-with-api/src/store/reducers/getDataReducer.jsx
@@ -1,0 +1,21 @@
+import { FETCH_DATA_START } from '../actions';
+
+const initialState = {
+    isFetching: false,
+    error: '',
+    dataArray: [],
+};
+
+export const dataFetchReducer = (state = initialState, action) => {
+    switch (action.type) {
+        case FETCH_DATA_START:
+            return {
+                ...state,
+                isFetching: true,
+                error: '',
+                dataArray: action.payload,
+            };
+        default:
+            return state;
+    }
+};
