+++ react-redux-with-api/src/store/reducers/index.js
@@ -1,25 +1,6 @@
 import { combineReducers } from 'redux';
-import { FETCH_DATA_START } from '../actions';
-const initialState = {
-    isFetching: false,
-    error: '',
-    dataArray: [],
-};
 
-export const dataFetchReducer = (state = initialState, action) => {
-    switch (action.type) {
-        case FETCH_DATA_START:
-            return {
-                ...state,
-                isFetching: true,
-                error: '',
-                dataArray: action.payload,
-            };
-        default:
-            return state;
-    }
-};
-
+import { dataFetchReducer } from './getDataReducer';
 export default combineReducers({
     dataFetchReducer,
 });
